requirements.txt
altair==4.2.0
anyio==3.3.1
appnope==0.1.2
argon2-cffi==21.1.0
astor==0.8.1
asyncio==3.4.3
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
cachetools==5.0.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cpp2python==0.2.0
cycler==0.10.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
dnspython==2.2.1
entrypoints==0.3
Flask==2.0.1
Flask-PyMongo==2.3.0
fyers-apiv2==2.0.4
get-nse-daily==0.2
gevent==21.8.0
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.11.2
imutils==0.5.4
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.3
jupyter-core==4.8.1
jupyter-server==1.11.0
jupyterlab==3.1.13
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.1
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lxml==4.6.3
MarkupSafe==2.0.1
matlab==0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
MouseInfo==0.1.3
mysql-connector==2.2.9
mysql-connector-python==8.0.26
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
nsepy==0.8
numpy==1.21.2
opencv-python==4.5.4.58
packaging==21.3
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
plotly==5.6.0
plotly-express==0.4.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.0
ptyprocess==0.7.0
pyaes==1.6.1
pyarrow==7.0.0
pyasn1==0.4.8
PyAutoGUI==0.9.53
pycparser==2.20
pydeck==0.7.1
PyGetWindow==0.0.9
Pygments==2.10.0
PyJWT==2.1.0
pymongo==4.1.0
Pympler==1.0.1
PyMsgBox==1.0.9
pyobjc==7.3
pyobjc-core==7.3
pyobjc-framework-Accessibility==7.3
pyobjc-framework-Accounts==7.3
pyobjc-framework-AddressBook==7.3
pyobjc-framework-AdServices==7.3
pyobjc-framework-AdSupport==7.3
pyobjc-framework-AppleScriptKit==7.3
pyobjc-framework-AppleScriptObjC==7.3
pyobjc-framework-ApplicationServices==7.3
pyobjc-framework-AppTrackingTransparency==7.3
pyobjc-framework-AuthenticationServices==7.3
pyobjc-framework-AutomaticAssessmentConfiguration==7.3
pyobjc-framework-Automator==7.3
pyobjc-framework-AVFoundation==7.3
pyobjc-framework-AVKit==7.3
pyobjc-framework-BusinessChat==7.3
pyobjc-framework-CalendarStore==7.3
pyobjc-framework-CallKit==7.3
pyobjc-framework-CFNetwork==7.3
pyobjc-framework-ClassKit==7.3
pyobjc-framework-CloudKit==7.3
pyobjc-framework-Cocoa==7.3
pyobjc-framework-Collaboration==7.3
pyobjc-framework-ColorSync==7.3
pyobjc-framework-Contacts==7.3
pyobjc-framework-ContactsUI==7.3
pyobjc-framework-CoreAudio==7.3
pyobjc-framework-CoreAudioKit==7.3
pyobjc-framework-CoreBluetooth==7.3
pyobjc-framework-CoreData==7.3
pyobjc-framework-CoreHaptics==7.3
pyobjc-framework-CoreLocation==7.3
pyobjc-framework-CoreMedia==7.3
pyobjc-framework-CoreMediaIO==7.3
pyobjc-framework-CoreMIDI==7.3
pyobjc-framework-CoreML==7.3
pyobjc-framework-CoreMotion==7.3
pyobjc-framework-CoreServices==7.3
pyobjc-framework-CoreSpotlight==7.3
pyobjc-framework-CoreText==7.3
pyobjc-framework-CoreWLAN==7.3
pyobjc-framework-CryptoTokenKit==7.3
pyobjc-framework-DeviceCheck==7.3
pyobjc-framework-DictionaryServices==7.3
pyobjc-framework-DiscRecording==7.3
pyobjc-framework-DiscRecordingUI==7.3
pyobjc-framework-DiskArbitration==7.3
pyobjc-framework-DVDPlayback==7.3
pyobjc-framework-EventKit==7.3
pyobjc-framework-ExceptionHandling==7.3
pyobjc-framework-ExecutionPolicy==7.3
pyobjc-framework-ExternalAccessory==7.3
pyobjc-framework-FileProvider==7.3
pyobjc-framework-FileProviderUI==7.3
pyobjc-framework-FinderSync==7.3
pyobjc-framework-FSEvents==7.3
pyobjc-framework-GameCenter==7.3
pyobjc-framework-GameController==7.3
pyobjc-framework-GameKit==7.3
pyobjc-framework-GameplayKit==7.3
pyobjc-framework-ImageCaptureCore==7.3
pyobjc-framework-IMServicePlugIn==7.3
pyobjc-framework-InputMethodKit==7.3
pyobjc-framework-InstallerPlugins==7.3
pyobjc-framework-InstantMessage==7.3
pyobjc-framework-Intents==7.3
pyobjc-framework-IOSurface==7.3
pyobjc-framework-iTunesLibrary==7.3
pyobjc-framework-KernelManagement==7.3
pyobjc-framework-LatentSemanticMapping==7.3
pyobjc-framework-LaunchServices==7.3
pyobjc-framework-libdispatch==7.3
pyobjc-framework-LinkPresentation==7.3
pyobjc-framework-LocalAuthentication==7.3
pyobjc-framework-MapKit==7.3
pyobjc-framework-MediaAccessibility==7.3
pyobjc-framework-MediaLibrary==7.3
pyobjc-framework-MediaPlayer==7.3
pyobjc-framework-MediaToolbox==7.3
pyobjc-framework-Metal==7.3
pyobjc-framework-MetalKit==7.3
pyobjc-framework-MetalPerformanceShaders==7.3
pyobjc-framework-MetalPerformanceShadersGraph==7.3
pyobjc-framework-MLCompute==7.3
pyobjc-framework-ModelIO==7.3
pyobjc-framework-MultipeerConnectivity==7.3
pyobjc-framework-NaturalLanguage==7.3
pyobjc-framework-NetFS==7.3
pyobjc-framework-Network==7.3
pyobjc-framework-NetworkExtension==7.3
pyobjc-framework-NotificationCenter==7.3
pyobjc-framework-OpenDirectory==7.3
pyobjc-framework-OSAKit==7.3
pyobjc-framework-OSLog==7.3
pyobjc-framework-PassKit==7.3
pyobjc-framework-PencilKit==7.3
pyobjc-framework-Photos==7.3
pyobjc-framework-PhotosUI==7.3
pyobjc-framework-PreferencePanes==7.3
pyobjc-framework-PushKit==7.3
pyobjc-framework-Quartz==7.3
pyobjc-framework-QuickLookThumbnailing==7.3
pyobjc-framework-ReplayKit==7.3
pyobjc-framework-SafariServices==7.3
pyobjc-framework-SceneKit==7.3
pyobjc-framework-ScreenSaver==7.3
pyobjc-framework-ScreenTime==7.3
pyobjc-framework-ScriptingBridge==7.3
pyobjc-framework-SearchKit==7.3
pyobjc-framework-Security==7.3
pyobjc-framework-SecurityFoundation==7.3
pyobjc-framework-SecurityInterface==7.3
pyobjc-framework-ServiceManagement==7.3
pyobjc-framework-Social==7.3
pyobjc-framework-SoundAnalysis==7.3
pyobjc-framework-Speech==7.3
pyobjc-framework-SpriteKit==7.3
pyobjc-framework-StoreKit==7.3
pyobjc-framework-SyncServices==7.3
pyobjc-framework-SystemConfiguration==7.3
pyobjc-framework-SystemExtensions==7.3
pyobjc-framework-UniformTypeIdentifiers==7.3
pyobjc-framework-UserNotifications==7.3
pyobjc-framework-UserNotificationsUI==7.3
pyobjc-framework-VideoSubscriberAccount==7.3
pyobjc-framework-VideoToolbox==7.3
pyobjc-framework-Virtualization==7.3
pyobjc-framework-Vision==7.3
pyobjc-framework-WebKit==7.3
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
pyrsistent==0.18.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.3
pyzmq==22.3.0
requests==2.25.1
requests-unixsocket==0.2.0
rsa==4.7.2
rubicon-objc==0.4.1
scipy==1.7.3
selenium==3.141.0
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smartapi-python==1.2.9
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.2.1
statsmodels==0.13.2
streamlit==1.7.0
telebot==0.0.4
Telethon==1.23.0
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==5.1.0
truedata-ws==4.1.0
twilio==7.8.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.6
validators==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.2.1
websockets==8.1
Werkzeug==2.0.1
widgetsnbextension==3.5.2
wikipedia==1.4.0
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
